Reorganize parser helpers
