+++ lib/parser/converter/converter.js
@@ -3,7 +3,7 @@
 const { VisitorOption } = require("estraverse");
 
 const { BaseVisitor } = require("./base-visitor");
-const { mix } = require("./mixin-builder");
+const { mix } = require("./helpers/mixin-builder");
 const { ExpressionParserMixin } = require("./expression-parser");
 const { ScriptBlockParserMixin } = require("./script-block-parser");
 const { PositionUpdaterMixin } = require("./position-updater");
