+++ lib/parser/converter/expression-parser.js
@@ -1,8 +1,6 @@
 "use strict";
 
-const { tokTypes, Token } = require("acorn");
-
-const { get_parser } = require("./espree");
+const { parseExpressionAt } = require("./helpers//espree");
 const { EXPRESSION_NODES } = require("../keys");
 
 const CODE = Symbol("ExpressionParser.code");
@@ -14,35 +12,6 @@ const PARSE_EXPRESSION = Symbol("ExpressionParser.parse_expression()");
 const EXPRESSION_NODE_ENTER = EXPRESSION_NODES.filter(type => type !== "Let")
   .map(type => `${type} > .expression`)
   .join(", ");
-
-// eslint-disable-next-line max-params
-function parseExpressionAt(code, base_options, start, end, assignable) {
-  const Parser = get_parser(start);
-  const parser = new Parser(base_options, code.slice(0, end));
-
-  parser.nextToken();
-
-  const destructuring_errors = {
-    shorthandAssign: -1,
-    trailingComma: -1,
-    parenthesizedAssign: -1,
-    parenthesizedBind: -1,
-    doubleProto: -1
-  };
-
-  let expression = parser.parseExpression(undefined, destructuring_errors);
-  if (assignable) {
-    expression = parser.toAssignable(expression, false, destructuring_errors);
-    parser.checkLVal(expression);
-  }
-
-  // espree holds closing curly braces hostage until the next token
-  parser.finishToken(tokTypes.eof);
-  parser.options.onToken(new Token(parser));
-
-  const { tokens, comments } = parser.get_extras();
-  return { expression, tokens, comments };
-}
 
 // eslint-disable-next-line max-lines-per-function
 function ExpressionParserMixin(SuperClass) {
