+++ lib/parser/converter/helpers/espree.js
@@ -34,4 +34,33 @@ function get_parser(position) {
   return acorn.Parser.extend(svelte(position), espree());
 }
 
-module.exports = { get_parser };
+// eslint-disable-next-line max-params
+function parseExpressionAt(code, base_options, start, end, assignable) {
+  const Parser = get_parser(start);
+  const parser = new Parser(base_options, code.slice(0, end));
+
+  parser.nextToken();
+
+  const destructuring_errors = {
+    shorthandAssign: -1,
+    trailingComma: -1,
+    parenthesizedAssign: -1,
+    parenthesizedBind: -1,
+    doubleProto: -1
+  };
+
+  let expression = parser.parseExpression(undefined, destructuring_errors);
+  if (assignable) {
+    expression = parser.toAssignable(expression, false, destructuring_errors);
+    parser.checkLVal(expression);
+  }
+
+  // espree holds closing curly braces hostage until the next token
+  parser.finishToken(acorn.tokTypes.eof);
+  parser.options.onToken(new acorn.Token(parser));
+
+  const { tokens, comments } = parser.get_extras();
+  return { expression, tokens, comments };
+}
+
+module.exports = { get_parser, parseExpressionAt };
