+++ lib/parser/converter/helpers/template-lexer.js
