+++ lib/parser/converter/helpers/token-mappings.js
