+++ lib/parser/converter/script-block-parser.js
@@ -1,6 +1,6 @@
 "use strict";
 
-const { get_parser } = require("./espree");
+const { get_parser } = require("./helpers/espree");
 
 const CODE = Symbol("ScriptBlockParser.code");
 const PARSER_OPTIONS = Symbol("ScriptBlockParser.parser_options");
