+++ lib/parser/converter/tag-tokenizer.js
@@ -3,14 +3,14 @@
 const {
   template_lexer,
   TOKEN_TYPES: LEXER_TOKEN_TYPES
-} = require("./template-lexer");
+} = require("./helpers/template-lexer");
 
 const {
   SKIP_AUTO_PUSH,
   ESPREE_TOKEN_TYPES,
   TOKEN_MAPPINGS,
   IDENTIFIERS
-} = require("./token-mappings");
+} = require("./helpers/token-mappings");
 
 const IMPORTED_COMPONENT_REFERENCE = "InlineComponent[name=/^[A-Z]/]";
